+++ src/main/kotlin/io/github/jamalam/pinguino/extensions/moderation/ModerationExtension.kt
@@ -51,7 +51,6 @@ import kotlinx.datetime.Clock
 import kotlinx.datetime.TimeZone
 import kotlinx.datetime.plus
 import java.util.*
-import kotlin.error
 import kotlin.time.Duration
 import kotlin.time.ExperimentalTime