+++ src/main/kotlin/io/github/jamalam/pinguino/extensions/user/UserUtilityExtension.kt
@@ -50,7 +50,6 @@ import io.github.jamalam.pinguino.util.*
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.toList
 import java.util.*
-import kotlin.error
 
 /**
  * @author  Jamalam360